refactor(api): move publish in batch route to src
