+++ api/src/certification/session-management/domain/models/SessionPublicationBatchResult.js
