+++ api/tests/certification/session-management/acceptance/application/session-controller-publish-session-in-batch_test.js
@@ -3,7 +3,7 @@ import {
   databaseBuilder,
   expect,
   generateValidRequestAuthorizationHeader,
-} from '../../../test-helper.js';
+} from '../../../../test-helper.js';
 
 describe('POST /api/admin/sessions/publish-in-batch', function () {
   let server;
