+++ api/tests/certification/session-management/unit/application/session-publication-controller_test.js
@@ -1,7 +1,10 @@
 import { sessionPublicationController } from '../../../../../src/certification/session-management/application/session-publication-controller.js';
+import { SessionPublicationBatchResult } from '../../../../../src/certification/session-management/domain/models/SessionPublicationBatchResult.js';
 import { usecases } from '../../../../../src/certification/session-management/domain/usecases/index.js';
+import { SessionPublicationBatchError } from '../../../../../src/shared/application/http-errors.js';
 import { getI18n } from '../../../../../src/shared/infrastructure/i18n/i18n.js';
-import { expect, hFake, sinon } from '../../../../test-helper.js';
+import { logger } from '../../../../../src/shared/infrastructure/utils/logger.js';
+import { catchErr, expect, hFake, sinon } from '../../../../test-helper.js';
 
 describe('Certification | Session-management | Unit | Application | Controller | Session Publication', function () {
   describe('#publish', function () {
@@ -73,6 +76,108 @@ describe('Certification | Session-management | Unit | Application | Controller |
 
       // then
       expect(response).to.equal(serializedSession);
+    });
+  });
+
+  describe('#publishInBatch', function () {
+    it('returns 204 when no error occurred', async function () {
+      // given
+      const i18n = getI18n();
+
+      const request = {
+        i18n,
+        payload: {
+          data: {
+            attributes: {
+              ids: ['sessionId1', 'sessionId2'],
+            },
+          },
+        },
+      };
+      sinon
+        .stub(usecases, 'publishSessionsInBatch')
+        .withArgs({
+          sessionIds: ['sessionId1', 'sessionId2'],
+          i18n,
+        })
+        .resolves(new SessionPublicationBatchResult('batchId'));
+
+      // when
+      const response = await sessionPublicationController.publishInBatch(request, hFake);
+
+      // then
+      expect(response.statusCode).to.equal(204);
+    });
+
+    it('logs errors when errors occur', async function () {
+      // given
+      const i18n = getI18n();
+      const result = new SessionPublicationBatchResult('batchId');
+      result.addPublicationError('sessionId1', new Error('an error'));
+      result.addPublicationError('sessionId2', new Error('another error'));
+
+      const request = {
+        i18n,
+        payload: {
+          data: {
+            attributes: {
+              ids: ['sessionId1', 'sessionId2'],
+            },
+          },
+        },
+      };
+      sinon.stub(usecases, 'publishSessionsInBatch').resolves(result);
+      sinon.stub(logger, 'warn');
+
+      // when
+      await catchErr(sessionPublicationController.publishInBatch)(request, hFake);
+
+      // then
+      expect(logger.warn).to.have.been.calledWithExactly(
+        'One or more error occurred when publishing session in batch batchId',
+      );
+
+      expect(logger.warn).to.have.been.calledWithExactly(
+        {
+          batchId: 'batchId',
+          sessionId: 'sessionId1',
+        },
+        'an error',
+      );
+
+      expect(logger.warn).to.have.been.calledWithExactly(
+        {
+          batchId: 'batchId',
+          sessionId: 'sessionId2',
+        },
+        'another error',
+      );
+    });
+
+    it('returns the serialized batch id', async function () {
+      // given
+      const i18n = getI18n();
+      const result = new SessionPublicationBatchResult('batchId');
+      result.addPublicationError('sessionId1', new Error('an error'));
+
+      const request = {
+        i18n,
+        payload: {
+          data: {
+            attributes: {
+              ids: ['sessionId1', 'sessionId2'],
+            },
+          },
+        },
+      };
+      sinon.stub(usecases, 'publishSessionsInBatch').resolves(result);
+      sinon.stub(logger, 'warn');
+
+      // when
+      const error = await catchErr(sessionPublicationController.publishInBatch)(request, hFake);
+
+      // then
+      expect(error).to.be.an.instanceof(SessionPublicationBatchError);
     });
   });
 });
